Move all possible code to a models package
