+++ GTA SA Chaos.sln
@@ -3,7 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GTA SA Chaos", "GTA SA Chaos.csproj", "{91106072-8642-4B48-AABE-F21A254FA788}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GtaSaChaos.Forms", "GtaSaChaos.Forms\GtaSaChaos.Forms.csproj", "{91106072-8642-4B48-AABE-F21A254FA788}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UI", "UI", "{C9DF719E-5EB3-4AA1-A0F6-6AC0E09BAD16}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Models", "Models", "{43294186-B53A-4338-839A-1DBC3343A6F2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GtaSaChaos.Models", "GtaSaChaos.Models\GtaSaChaos.Models.csproj", "{D0887711-4CE7-4085-8AD6-CD46C309759B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,9 +33,25 @@ Global
 		{91106072-8642-4B48-AABE-F21A254FA788}.Release|x64.Build.0 = Release|Any CPU
 		{91106072-8642-4B48-AABE-F21A254FA788}.Release|x86.ActiveCfg = Release|Any CPU
 		{91106072-8642-4B48-AABE-F21A254FA788}.Release|x86.Build.0 = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|x64.Build.0 = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Debug|x86.Build.0 = Debug|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|x64.ActiveCfg = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|x64.Build.0 = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|x86.ActiveCfg = Release|Any CPU
+		{D0887711-4CE7-4085-8AD6-CD46C309759B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{91106072-8642-4B48-AABE-F21A254FA788} = {C9DF719E-5EB3-4AA1-A0F6-6AC0E09BAD16}
+		{D0887711-4CE7-4085-8AD6-CD46C309759B} = {43294186-B53A-4338-839A-1DBC3343A6F2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6150737D-6AAD-4696-BFBE-5301699F4DB3}
