+++ GtaSaChaos.Forms/Form1.cs
@@ -1,16 +1,17 @@
 // Copyright (c) 2019 Lordmau5
-using GTA_SA_Chaos.effects;
-using GTA_SA_Chaos.util;
-using Newtonsoft.Json;
+
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Windows.Forms;
+using GtaSaChaos.Models.Effects;
+using GtaSaChaos.Models.Effects.@abstract;
+using GtaSaChaos.Models.Utils;
+using Newtonsoft.Json;
 
-namespace GTA_SA_Chaos
+namespace GtaSaChaos.Forms
 {
     public partial class Form1 : Form
     {
