+++ GtaSaChaos.Forms/Program.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) 2019 Lordmau5
+
 using System;
 using System.Windows.Forms;
 
-namespace GTA_SA_Chaos
+namespace GtaSaChaos.Forms
 {
     internal static class Program
     {
