+++ GtaSaChaos.Models/GtaSaChaos.Models.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Utils\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
+    <PackageReference Include="TwitchLib.Api" Version="3.0.6-CI-20190905-060235" />
+    <PackageReference Include="TwitchLib.Client" Version="3.0.6-CI-20190916-073325" />
+  </ItemGroup>
+
+</Project>
